refactor: moved SideBarLink and Theme Button components inside the SideBar component folder and removed export default.
