+++ src/components/SideBar/SideBarLink/SideBarLink.spec.tsx
@@ -1,7 +1,17 @@
 import { screen } from '@testing-library/react';
 import { FiHome } from 'react-icons/fi';
-import SideBarLink from '.';
-import { renderWithTheme } from '../../tests/helpers/renderWithTheme';
+import { SideBarLink } from '.';
+import { renderWithTheme } from '../../../tests/helpers/renderWithTheme';
+
+jest.mock('next/router', () => {
+  return {
+    useRouter() {
+      return {
+        asPath: '/'
+      }
+    }
+  }
+});
 
 describe('SideBarLink component', () => {
   it('should render a non-activable link correctly', async () => {
@@ -11,7 +21,7 @@ describe('SideBarLink component', () => {
   });
 
   it('should render an activable link that is active correctly', async () => {
-    renderWithTheme(<SideBarLink title="Test" icon={FiHome} isActivable={true} isActive={true} />, { themeName: 'light' })
+    renderWithTheme(<SideBarLink title="Test" icon={FiHome} isActivable={true} href="/" />, { themeName: 'light' })
 
     expect(screen.getByTestId('active-bar')).toBeInTheDocument();
   });
